refactored rust code in tauri app
