+++ src-tauri/src/commands/logs.rs
@@ -1,0 +1,73 @@
+use crate::types::LogEntry;
+use std::fs::{self, OpenOptions};
+use std::io::{BufRead, BufReader, Write};
+use std::path::{Path, PathBuf};
+
+fn get_log_file_path() -> PathBuf {
+    let biovault_home = std::env::var("BIOVAULT_HOME").unwrap_or_else(|_| {
+        let home_dir = dirs::home_dir().unwrap();
+        dirs::desktop_dir()
+            .unwrap_or_else(|| home_dir.join("Desktop"))
+            .join("BioVault")
+            .to_string_lossy()
+            .to_string()
+    });
+    Path::new(&biovault_home).join("desktop_commands.log")
+}
+
+#[allow(dead_code)]
+pub fn append_log(entry: &LogEntry) -> Result<(), String> {
+    let log_path = get_log_file_path();
+
+    // Ensure directory exists
+    if let Some(parent) = log_path.parent() {
+        fs::create_dir_all(parent).map_err(|e| format!("Failed to create log directory: {}", e))?;
+    }
+
+    let json_line = serde_json::to_string(entry)
+        .map_err(|e| format!("Failed to serialize log entry: {}", e))?;
+
+    let mut file = OpenOptions::new()
+        .create(true)
+        .append(true)
+        .open(&log_path)
+        .map_err(|e| format!("Failed to open log file: {}", e))?;
+
+    writeln!(file, "{}", json_line).map_err(|e| format!("Failed to write to log file: {}", e))?;
+
+    Ok(())
+}
+
+#[tauri::command]
+pub fn get_command_logs() -> Result<Vec<LogEntry>, String> {
+    let log_path = get_log_file_path();
+
+    if !log_path.exists() {
+        return Ok(Vec::new());
+    }
+
+    let file =
+        std::fs::File::open(&log_path).map_err(|e| format!("Failed to open log file: {}", e))?;
+
+    let reader = BufReader::new(file);
+    let mut logs = Vec::new();
+
+    for line_str in reader.lines().map_while(Result::ok) {
+        if let Ok(entry) = serde_json::from_str::<LogEntry>(&line_str) {
+            logs.push(entry);
+        }
+    }
+
+    Ok(logs)
+}
+
+#[tauri::command]
+pub fn clear_command_logs() -> Result<(), String> {
+    let log_path = get_log_file_path();
+
+    if log_path.exists() {
+        fs::remove_file(&log_path).map_err(|e| format!("Failed to delete log file: {}", e))?;
+    }
+
+    Ok(())
+}
