+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,10 @@
+pub mod dependencies;
+pub mod files;
+pub mod jupyter;
+pub mod logs;
+pub mod messages;
+pub mod participants;
+pub mod projects;
+pub mod runs;
+pub mod settings;
+pub mod syftbox;
